Changed types to double and currently making modifications to console

+++ src/starter/Coordinates.java
@@ -1,10 +1,10 @@
 package starter;
 
 public class Coordinates {
-	public int x;
-	public int y;
+	public double x;
+	public double y;
 	
-	public Coordinates(int horizontal, int vertical)
+	public Coordinates(double horizontal, double vertical)
 	{
 		x = horizontal;
 		y = vertical;
@@ -12,20 +12,20 @@ public class Coordinates {
 	
 	public Coordinates() {}
 	
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 	
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 	
-	public int getX() 
+	public double getX() 
 	{
 		return x;
 	}
 	
-	public int getY() 
+	public double getY() 
 	{
 		return y;
 	}
